Improved handling multiple commits from the same checkout
